+++ Helper/EnemyDefinition.cs
@@ -7,11 +7,11 @@ using System.Threading.Tasks;
 public class EnemyDefinition
 {
     public string Type { get; set; }
-    public int X { get; set; }
-    public int Y { get; set; }
+    public float X { get; set; }
+    public float Y { get; set; }
     public int count { get; set; } // used for wallmaster manager
 
-    public EnemyDefinition(string type, int x, int y, int count = 1)
+    public EnemyDefinition(string type, float x, float y, int count = 1)
     {
         Type = type;
         X = x;
